Update docker-compose.yml and nginx.conf
This commit updates the docker-compose.yml file to remove the port mapping for Django and add a port mapping for HTTPS. It also adds the 'transcendence' network to the nginx service. Additionally, the nginx.conf file is modified to include a location block for the '/api' path, which proxies requests to the Django service.